job_administration_public/requirements.txt
easygui==0.98.3
Flask==3.1.2
fpdf2==2.8.5
numpy==2.3.5
selenium==4.39.0
